+++ person.py
@@ -4,6 +4,7 @@ from resource import fallback_answer,answer, pers_step, pers_zag,pers_sprav
 
 
 def person(event,step,place, status=None, id_zag=None):
+    event['state']['context']='quest_begin'
     intents=event['request'].get('nlu',{}).get('intents')
     # if step==0:
 #         Выбираем ИД загадки
@@ -14,30 +15,30 @@ def person(event,step,place, status=None, id_zag=None):
         if step==0:
             return make_response(text=param[0],tts=param[1],buttons=[
             button('Повтори', hide=True), button('Жар-птицу', hide=True),button('Я не знаю', hide=True),],step=step+1, place=place,
-            status=param[2], event=event, context='quest_begin')
+            status=param[2], event=event)
         elif step==1:
             if 'gift' in intents:
                 return make_response(text=param[0],tts=param[1],buttons=[
                     button('Да', hide=True),button('Нет', hide=True)],step=step+2, place=place,
-                                    status=param[2], event=event, context='quest_begin')
+                                    status=param[2], event=event)
             elif 'povtor'in intents or "YANDEX.REPEAT" in intents:
                 step=0
                 param=text_to_resp(pers_step,place,step)
                 return make_response(text=param[0],tts=param[1],buttons=[
             button('Повтори', hide=True), button('Жар-птицу', hide=True),button('Я не знаю', hide=True),],step=step+1, place=place,
-            status=param[2], event=event, context='quest_begin')                    
+            status=param[2], event=event)                    
             
             else:
                 step=step+1
                 param=text_to_resp(pers_step,place,step)
                 return make_response(text=param[0],tts=param[1], buttons=[
                     button('Да', hide=True),button('Нет', hide=True)],step=step+1, place=place,
-                                    status=param[2], event=event, context='quest_begin')
+                                    status=param[2], event=event)
             
 # Обработка ответа "да"
         elif 'answer_da' in intents or 'YANDEX.CONFIRM' in intents:
             if step ==4 or step ==6:
-                return make_response(text='Говори!', step=step,place=place,status=status, event=event, context='quest_begin')
+                return make_response(text='Говори!', step=step,place=place,status=status, event=event)
             else:
                 if step==3:
                     param=text_to_resp(pers_zag,pers_sprav[place][0],0)
@@ -45,7 +46,7 @@ def person(event,step,place, status=None, id_zag=None):
                     param=text_to_resp(pers_zag,pers_sprav[place][1],0)
                 return make_response(text=param[0],tts=param[1],buttons=[
                     button('Повтори', hide=True),button('Я не знаю', hide=True),button('Я знаю', hide=True)], 
-                                     step=step+1, place=place,status=param[2], event=event, context='quest_begin')
+                                     step=step+1, place=place,status=param[2], event=event)
 #  Обработка ответа "Нет"
 
         elif 'net' in intents or 'YANDEX.REJECT' in intents:
@@ -59,13 +60,13 @@ def person(event,step,place, status=None, id_zag=None):
                         param=text_to_resp(fallback_answer,place,2)
                         return make_response(text=param[0],tts=param[1],buttons=[
                     button('Да', hide=True),button('Нет', hide=True)], step=step, place=place,
-                                    status=param[2], event=event, context='quest_begin')
+                                    status=param[2], event=event)
             elif step == 5:
                 param=text_to_resp(pers_zag,pers_sprav[place][1],1,obj_3=True)           
                 return make_response(text=param[0],tts=param[1],buttons=[
                     button('Купол', hide=True),button('Звонница', hide=True),
                     button('Собор', hide=True),], step=0, place=place,
-                    status=param[2],place_next=param[3], event=event, context='quest_begin')    
+                    status=param[2],place_next=param[3], event=event)    
             elif step == 4:
                 return help_4_zagadka (event,pers_sprav[place][0])
             elif step==6:
@@ -94,19 +95,19 @@ def person(event,step,place, status=None, id_zag=None):
                     param=text_to_resp(pers_zag,pers_sprav[place][1],0)
                     buttons=[button('Повтори', hide=True),button('Я не знаю', hide=True),button('Я знаю', hide=True)]
                 return make_response(text=param[0],tts=param[1],buttons=buttons,
-                                     step=step, place=place,status=param[2], event=event, context='quest_begin')
+                                     step=step, place=place,status=param[2], event=event)
 # обработка правильных ответов
         elif answer[place][pers_sprav[place][0]] in intents and step==4:
                 param=text_to_resp(pers_zag,pers_sprav[place][0],1)
                 return make_response(text=param[0],tts=param[1],buttons=[
                     button('Да', hide=True),button('Нет', hide=True)], step=step+1, place=place,
-                                    status=param[2], event=event, context='quest_begin')
+                                    status=param[2], event=event)
         elif answer[place][pers_sprav[place][1]] in intents and step==6:
                 param=text_to_resp(pers_zag,pers_sprav[place][1],1,obj_3=True)
                 return make_response(text=param[0],tts=param[1],buttons=[
             button('Купол', hide=True),button('Звонница', hide=True),
             button('Собор', hide=True),], step=0, place=place,
-                                    status=param[2],place_next=param[3], event=event, context='quest_begin')
+                                    status=param[2],place_next=param[3], event=event)
 # Обработка else
 
         else:
@@ -118,7 +119,7 @@ def person(event,step,place, status=None, id_zag=None):
                     param=text_to_resp(fallback_answer,place,0)
                     return make_response(text=param[0],tts=param[1],buttons=[
                     button('Да', hide=True),button('Нет', hide=True)], step=step, place=place,
-                                    status=param[2], event=event, context='quest_begin')
+                                    status=param[2], event=event)
             else:
                 if step==4:
                     return help_4_zagadka (event,pers_sprav[place][0])
