+++ sights.py
@@ -2,7 +2,7 @@ sights = {
 
   'kupol': {
     'name': "Па́мятник Тысячеле́тию Росси́и",
-    'to_tip_name': 'бронзовому куполу тысячилетнему, что в середине земель кремлёвских',
+    'to_tip_name': 'Кинул Садко клубок и покатился он к бронзовому куполу тысячелетнему, что в середине земель кремлёвских. Следуй и ты за ним путник, да смотри не отставай! А пока ты идешь я могу тебе про него рассказать. Хочешь?',
     'location': {
       'lat': 58.52114,
       'lon': 31.27527,
